prevent sys.exit in lambdas with SecurityManager
