+++ starport-lambda/src/main/scala/com/krux/starport/lambda/SubmitHandler.scala
@@ -1,31 +1,19 @@
 package com.krux.starport.lambda
 
 import java.io.ByteArrayOutputStream
-import java.security.Permission
 
 import com.amazonaws.services.lambda.runtime.{Context, RequestHandler}
 import com.krux.starport.db.tool.SubmitPipeline
 import org.slf4j.LoggerFactory
 
-
-sealed case class ExitException(status: Int) extends SecurityException("System.exit() called but ignored") {
-}
-
-sealed class NoExitSecurityManager extends SecurityManager {
-  override def checkPermission(perm: Permission): Unit = {}
-
-  override def checkPermission(perm: Permission, context: Object): Unit = {}
-
-  override def checkExit(status: Int): Unit = {
-    super.checkExit(status)
-    throw ExitException(status)
-  }
-}
-
 /**
   * Lambda to invoke the com.krux.starport.db.tool.SubmitPipeline util remotely.
   *
+  * Property configuration for lambda (pass via JAVA_TOOL_OPTIONS env var):
+  *
   * -Dstarport.config.url=s3://{your-bucket}/starport/starport.conf
+  * -Dlogger.root.level=DEBUG|INFO|...
+  * -Dexecution.context=lambda
   *
   */
 class SubmitHandler extends RequestHandler[SubmitRequest, SubmitResponse] {
@@ -35,15 +23,12 @@ class SubmitHandler extends RequestHandler[SubmitRequest, SubmitResponse] {
   def handleRequest(input: SubmitRequest, context: Context): SubmitResponse = {
     val outCapture = new ByteArrayOutputStream
     val errCapture = new ByteArrayOutputStream
+    //      Console.withOut(outCapture) {
+    //        Console.withErr(errCapture) {
     logger.info("lambda invoked...")
     try {
-      Console.withOut(outCapture) {
-        Console.withErr(errCapture) {
-          SubmitPipeline.main(input.getArgs)
-        }
-      }
+      SubmitPipeline.main(input.getArgs)
     } catch {
-      case exit: ExitException => logger.warn("captured System.exit error", exit)
       case unknown: Throwable => logger.error("unhandled error", unknown)
     } finally {
       logger.info("lambda finished...")
